requirements.txt
aead==0.2
agate==1.7.1
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.9
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.1
bigquery-schema-generator==1.5.1
bleach==5.0.1
blinker==1.6.3
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver_autoinstaller==0.6.2
click==8.1.3
colorama==0.4.6
comm==0.1.4
coverage==7.3.1
cryptography==38.0.4
csv-diff==1.1
csvdiff==0.3.3
csvkit==1.1.1
dbfread==2.0.7
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
distlib==0.3.7
dnspython==2.4.2
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.12.4
Flask==3.0.0
fqdn==1.5.1
frozenlist==1.3.3
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
leather==0.3.4
lxml==4.9.3
macholib==1.16.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.1.0
nba-api==1.2
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
openai==0.27.2
openapi==1.1.0
openpyxl==3.1.2
outcome==1.2.0
packaging==22.0
pandas==1.5.2
pandasql==0.7.3
pandocfilters==1.5.0
parsedatetime==2.6
parso==0.8.3
peewee==3.16.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.11
platformdirs==3.10.0
pluggy==1.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
proto-plus==1.22.3
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.4.0
Pygments==2.13.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
pymongo==4.5.0
PyMySQL==1.0.2
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.4.2
pytest-cov==2.12.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python-slugify==8.0.1
PythonTurtle==0.3.2
pytimeparse==1.1.8
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.0
rand==0.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
secure-smtplib==0.1.1
selenium==4.13.0
Send2Trash==1.8.0
six==1.16.0
slack-sdk==3.21.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.48
stack-data==0.6.2
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
trio==0.22.2
trio-websocket==0.11.1
tweepy==4.14.0
uri-template==1.2.0
urllib3==1.26.13
virtualenv==20.24.5
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.0.0
widgetsnbextension==4.0.8
wsproto==1.2.0
xlrd==2.0.1
yarg==0.1.9
yarl==1.8.2
